Fix for empty string returning true during validation
Problem: passing the value `''` to `validate()` results in a return value of `true`.

Solution: remove a length check that was returning `true` for a length of 0.

NOTE: I was really trying to find a reason as to why you would want to return `true` for an empty string but I couldn't think of a reason. I concluded that this was possibly leftover code that was no longer needed. (`git blame` showed it being added over four years ago)

Thanks for publishing and maintaing a great helper library!